+++ app/services/forest_liana/resources_getter.rb
@@ -11,10 +11,9 @@ module ForestLiana
       @collection_name = ForestLiana.name_for(@resource)
       @collection = get_collection(@collection_name)
       @field_names_requested = field_names_requested
-      @tables_associated_to_relations_name = {}
       get_segment()
       compute_includes()
-      @search_query_builder = SearchQueryBuilder.new(@params, @includes, @tables_associated_to_relations_name, @collection)
+      @search_query_builder = SearchQueryBuilder.new(@params, @includes, @collection)
 
       prepare_query()
     end
@@ -40,13 +39,6 @@ module ForestLiana
 
     def compute_includes
       associations_has_one = ForestLiana::QueryHelper.get_one_associations(@resource)
-
-      includes = associations_has_one.each do |association|
-        if @tables_associated_to_relations_name[association.table_name].nil?
-          @tables_associated_to_relations_name[association.table_name] = []
-        end
-        @tables_associated_to_relations_name[association.table_name] << association.name
-      end
 
       includes = associations_has_one.map(&:name)
       includes_for_smart_search = []